move out manage host filters into their own component (#10749)
# Checklist for submitter

This removes the filtering UI from the manage hosts page. Currently it
keeps the same code (I felt it was too risky to move out the code AND do
a big rewrite) but I have other work still in progress where I try to
improve the code for filtering.

Basically this is the first step to making the code for creating and
maintaining filters a bit easier.

- [x] Manual QA for all new/changed functionality
