+++ frontend/pages/hosts/ManageHostsPage/_styles.scss
@@ -248,34 +248,6 @@
     margin-left: $pad-small;
   }
 
-  &__macsettings-dropdown {
-    width: 137px;
-
-    .Select-value {
-      display: flex;
-      align-items: center;
-
-      &::before {
-        position: relative;
-        content: url(../assets/images/icon-filter-v2-black-16x16@2x.png);
-        transform: scale(0.5);
-        left: -8px;
-        top: 4px;
-      }
-    }
-  }
-
-  &__labels-active-filter-wrap {
-    display: flex;
-    align-items: center;
-    margin-bottom: $pad-medium;
-    gap: $pad-small; // between multiple filter pills
-  }
-
-  &__policies-filter-pill {
-    margin-left: $pad-medium;
-  }
-
   &__enroll-hosts {
     padding: $pad-small;
     margin-right: $pad-small;
